implement bday_next command + structure format + lang
